Add a dummy serialVersionUID member to new serializable classes.
Classes that are serializable, in this case because they inherit from
serializable java.lang classes, must provide a class version marker,
which is used to validate the object against the class before unpacking.

Best practice is to make this a hash of the class signature, such as
generated by the 'serialver' tool. However, we don't expect any
serialization to happen within the applet, so for now we just set it
to 1L to silence warnings about the missing member, as a number of
other cortado classes were already doing.