Fix for Theora streams with pixel aspect ratio of zero or infinity.
Theora stream can have a PAR with numerator or denominator equal to zero.
This is a valid configuration; according to the specification it indicates
that the actual ratio is unknown and the player MAY assume 1:1, which is
what this patch does.

(Negative aspect ratio is still illegal and maybe even unrepresentable, so we
also catch that here.)